do not run all test every ime

+++ script/prepare_manageiq.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 CONTAINER_NAME=${CONTAINER_NAME-manageiq}
 TAG=${TAG-hammer-1-rc1}
+SUITE=${SUITE-default}
+if [[ $SUITE != "integration" ]]; then exit 0; fi
 
 docker rm -f $CONTAINER_NAME
 docker pull manageiq/manageiq:$TAG
